fix pull request issues
