--- radboudumc-template/templates/new/publication.html
@@ -1,42 +1,0 @@
-{% extends "base.html" %}
-{% block title %}{{ page.title }}{%endblock%}
-{% block content %}
-    {% include 'blocks/breadcrumbs.html' %}
-
-    <div class="container">
-    <h1 class="page-title">{{ page.title }}</h1>
-
-    <p class="lead">
-        {% for author in page.authors %}
-            {{ author }}{% if not loop.last %}, {% endif %}
-        {% endfor %}
-    </p>
-
-    <p>{{ page.pub_details }}</p>
-    <p>
-        {% if page.doi %}
-            <a href="{{ page.doi }}" class="btn btn-primary">DOI</a>
-        {% endif %}
-        {% if page.arxiv %}
-            <a href="{{ page.arxiv }}" class="btn btn-primary">arXiv</a>
-        {% endif %}
-        {% if page.pmid %}
-            <a href="{{ page.pmid }}" class="btn btn-primary">PMID</a>
-        {% endif %}
-        {% if page.urlweb %}
-            <a href="{{ page.urlweb }}" class="btn btn-primary">URL</a>
-        {% endif %}
-    </p>
-    <h2>Abstract</h2>
-    {{ page.content }}
-
-
-    {% if ENABLE_PUBLICATION_PDF_REQUEST and page.has_pdf == 'True' %}
-    <h2>Request PDF</h2>
-
-    <p>A pdf file of this publication is available for personal use. Enter your e-mail address in the box below and press the button. You will receive an e-mail message with a link to the pdf file. Note: for this service you will be redirected to our PDF-service at <a href="http://www.radboudimaging.nl">radboudimaging.nl</a>.
-
-    <form method="post" action="http://shared.radboudimaging.nl/mediawiki/publication/confirmemail.php?bibkey={{ page.bibkey }}&amp;ip=__USERIP__"><input type="text" name="email" size="50"><input type="submit" value="Send code by e-mail"></form>
-    {% endif %}
-    </div>
-{% endblock %}
