--- radboudumc-template/templates/new/publications.html
@@ -1,123 +1,0 @@
-{% extends "base.html" %}
-{% block title %}{{ page.title }}{%endblock%}
-{% block description %}{{ SITENAME }} publications overview.{% endblock %}
-{% block page_picture %}{{ SITEURL }}/{{SITE_PICTURE}}{% endblock %}
-{% block content %}
-    {% include 'blocks/breadcrumbs.html' %}
-    <div class="container">
-    <h1>{{ page.title }}</h1>
-
-
-
-    {% set list_years = [] %}
-    {% set list_pub_type = [] %}
-
-
-    {% if bibkeys_html %}
-    {% for bibkey in bibkeys_html.keys() %}
-        {# Getting list of years #}
-        {% if bibkeys_html[bibkey]['year'] not in list_years %}
-            {% if list_years.append(bibkeys_html[bibkey]['year']) %}{% endif %}
-        {% endif %}
-        {# Getting list of pub_type #}
-        {% if bibkeys_html[bibkey]['pub_type'] not in list_pub_type %}
-            {% if list_pub_type.append(bibkeys_html[bibkey]['pub_type']) %}{% endif %}
-        {% endif %}
-    {% endfor %}
-    {% endif %}
-    <style>
-      input[name="year-selection"]:checked + label {
-      font-weight: bold;
-      }
-      input[name="year-selection"]:hover + label {
-      font-weight: bold;
-      text-decoration: none;
-      }
-
-    </style>
-    <div class="btn-group btn-group-toggle" data-toggle="buttons">
-    {% set max_year = MAX_YEAR_PUB or 2012   %}
-    {% for cur_year in list_years|sort(reverse = True) if cur_year >= max_year  %}
-        {% if cur_year <= max_year %}
-            {% set cur_year = max_year|string + ' and before' %}
-        {% endif %}
-    <input type="radio" name="year-selection" data-year="{{ cur_year }}" id="input_{{ cur_year }}" autocomplete="off" style="display: none">
-    <label class="year-button btn btn-link" for="input_{{cur_year}}", style="color:grey">
-        {{ cur_year }}
-    </label>
-    {% endfor %}
-    </div>
- 
-    {# Inpress keys should go here. To include "Accepted or to appear" section #}
-
-    {# Lists the publications per year and organized by publication type #}
-    {% for cur_year in list_years|sort(reverse = True) %}
-        <div data-year="{{ cur_year }}" class="publication-year">
-            <h2>{{ cur_year }}</h2>
-            {% for pub_type in list_pub_type %}
-                {% set found = { 'flag': false } %}
-                {% for bibkey in bibkeys_html.keys() %}
-                    {% if bibkeys_html[bibkey]['pub_type'] == pub_type and bibkeys_html[bibkey]['year'] == cur_year %}
-                        {% if found['flag'] == false %}
-                            {% if pub_type == '@Article' %}
-                                <h3>{{ 'Papers in international journals' }}</h3>
-                            {% endif%}
-                            {% if pub_type == '@Preprint' %}
-                                <h3>{{ 'Preprints' }}</h3>
-                            {% endif%}
-                            {% if pub_type == '@InProceedings' %}
-                                <h3>{{ 'Papers in conference proceedings' }}</h3>
-                            {% endif%}
-                            {% if pub_type == '@Conference' %}
-                                <h3>{{ 'Abstracts' }}</h3>
-                            {% endif%}
-                            {% if pub_type in ['@PhdThesis', '@MastersThesis'] %}
-                                <h3>{{ 'Theses' }}</h3>
-                            {% endif%}
-                            {% if pub_type in ['@Book', '@Other'] %}
-                                <h3>{{ 'Other publications' }}</h3>
-                            {% endif%}
-                            <ul>
-                            {% if found.update({'flag': true}) %}{% endif %}
-                        {% endif %}
-                        {{ bibkeys_html[bibkey]['html'].replace('{filename}/pages', SITEURL).replace('.md', '/') }}
-                    {% endif %}
-                {% endfor %}
-                {% if found['flag'] == true %}
-                    </ul>
-                {% endif %}
-            {% endfor %}
-        </div>
-    {% endfor %}
-
-    </div>
-
-<script type="text/javascript">
-    var buttons = document.getElementsByName('year-selection');
-    var year_blocks = document.getElementsByClassName('publication-year');
-
-    for(var i = 0; i < buttons.length; i++) {
-        buttons[i].addEventListener('click', function(event) {
-        var year = this.dataset.year; 
-        var yearsplit = year.split(' ')
-        for(var b = 0; b < year_blocks.length; b++) {
-            if (yearsplit[yearsplit.length-1] == 'before'){
-                if(year_blocks[b].dataset.year <= parseInt(yearsplit[0])) {
-                    year_blocks[b].style.display = 'block';
-                } else {
-                    year_blocks[b].style.display = 'none';
-                }
-    
-            } else {
-                if(year_blocks[b].dataset.year == year) {
-                    year_blocks[b].style.display = 'block';
-                } else {
-                    year_blocks[b].style.display = 'none';
-                }
-            }
-      }
-    });
-  }
-  buttons[0].click()
-</script>
-{% endblock %}
