--- radboudumc-template/templates/newnew/publication-thesis.html
@@ -1,47 +1,0 @@
-{% extends "base.html" %}
-{% block title %}{{ page.title }}{%endblock%}
-{% block description %}{{ page.content | striptags | truncate(200)}}{% endblock %}
-{% block page_picture %}{{SITEURL}}/images/theses/{{ page.coverpng }}{% endblock %}
-{% block content %}
-    {% include 'blocks/breadcrumbs.html' %}
-
-    <div class="container">
-    <h1 class="page-title">{{ page.title }}</h1>
-
-    <p class="lead">
-        {% for author in page.authors %}
-            {{ author }}
-        {% endfor %}
-    </p>
-    <img src="{{SITEURL}}/images/theses/{{ page.coverpng }}">
-    {% if page.promotor %}
-        <p><b>Promotor: {{ page.promotor }}</b></p>
-    {% endif %}
-    {% if page.copromotor %}
-        <p><b>Copromotor: {{ page.copromotor }}</b></p>
-    {% endif %}
-    {% if page.school %}
-        <p><b>{{ page.school }}</b></p>
-    {% endif %}
-    {% if page.optmonth or page.year %}
-        <p><b>
-            {% if page.optmonth %}
-                {{page.optmonth.replace(',', '')}},
-            {% endif %}
-            {% if page.year %}
-                {{page.year}}
-            {% endif %}
-        </b></p>
-    {% endif %}
-
-    <p>{{ page.pub_details }}</p>
-    <p>
-        {% if page.urlweb %}
-            <a href="{{ page.urlweb }}" class="btn btn-primary">URL</a>
-        {% endif %}
-    </p>
-    <h2>Abstract</h2>
-    {{ page.content }}
-
-    </div>
-{% endblock %}
