--- radboudumc-template/templates/newnew/publication.html
@@ -1,9 +1,0 @@
-{% extends "base.html" %}
-{% block title %}{{ page.title }}{%endblock%}
-{% block content %}
-    {% include 'blocks/breadcrumbs.html' %}
-
-    <div class="container">
-	hoi
-    </div>
-{% endblock %}
