--- radboudumc-template/templates/newnew/publications-author.html
@@ -1,8 +1,0 @@
-{% extends "base.html" %}
-{% block title %}{{ page.title }}{%endblock%}
-{% block content %}
-    {% include 'blocks/breadcrumbs.html' %}
-    <div class="container">
-	hoi
-    </div>
-{% endblock %}
