+++ radboudumc-template/templates/thesis-gallery.html
@@ -8,20 +8,13 @@
     <div class="container">
     <h1>{{ page.title }}</h1>
 
-    {% set list_years = [] %}
-    {% set list_pub_type = [] %}
-
     {% set bib_keys = [] %}
+    <!-- get bibkeys for current group -->
     {% for person in pages|reverse %}
     {% if person.parent.url == 'members/' %}
-       {% set person_key = 'ja' %}  
-       {% if person.pub_name %}
-        {% set person_key = person.pub_name.replace(' ', '-') %}  
-       {% else %}
-        {% set person_key = person.name.replace(' ', '-') %}  
-       {% endif %}
-       {% if person_key in author_keys %}
-           {% for bibkey in author_keys[person_key]  %}
+       {% set person_key = person.name.replace(' ', '-') %}  
+       {% if person_key in AUTHOR_KEYS %}
+           {% for bibkey in AUTHOR_KEYS[person_key]  %}
             {% if bibkey not in bib_keys %}
                 {% set _ = bib_keys.append(bibkey) %}
             {% endif %}
@@ -30,21 +23,24 @@
     {% endif %}
     {% endfor %}
     
-    
+    {% set list_years = [] %}
+    {% set list_pub_type = [] %}
     {% if bib_keys %}
     {% for bibkey in bib_keys %}
         <!-- Getting list of years -->
-        {% if bib_items[bibkey]['type'] == 'phdthesis' and bib_items[bibkey]['year'] not in list_years %}
-            {% if list_years.append(bib_items[bibkey]['year']) %}{% endif %}
+        {% if BIB_ITEMS[bibkey]['type'] == 'phdthesis' and BIB_ITEMS[bibkey]['year'] not in list_years %}
+            {% if list_years.append(BIB_ITEMS[bibkey]['year']) %}{% endif %}
         {% endif %}
         <!-- Getting list of pub_type -->
-        {% if bib_items[bibkey]['type'] == 'phdthesis' and bib_items[bibkey]['type'] not in list_pub_type %}
-            {% if list_pub_type.append(bib_items[bibkey]['type']) %}{% endif %}
+        {% if BIB_ITEMS[bibkey]['type'] == 'phdthesis' and BIB_ITEMS[bibkey]['type'] not in list_pub_type %}
+            {% if list_pub_type.append(BIB_ITEMS[bibkey]['type']) %}{% endif %}
         {% endif %}
     {% endfor %}
     {% endif %}
+    
     <!-- Inpress keys should go here. To include "Accepted or to appear" section -->
     <p>Below is an overview of PhD theses written or supervised by members of the Diagnostic Image Analysis Group.</p>
+
     <!-- Lists the publications per year and organized by publication type -->
     <div class="row">
     {% set theses_without_cover = [] %}
@@ -52,16 +48,15 @@
             {% for pub_type in list_pub_type %}
                 {% set found = { 'flag': false } %}
                 {% for bibkey in bib_keys %}
-                    {% if bib_items[bibkey]['type'] == pub_type and bib_items[bibkey]['year'] == cur_year %}
-                    {% set coverpng = bib_items[bibkey]['coverpng'] %}
-                    {% set bibkey = bib_items[bibkey]['coverpng'].replace('.png', '').lower() %}
-                        {% if bib_items[bibkey]['cover_exists'] == 'True' %}
+                    {% if BIB_ITEMS[bibkey]['type'] == pub_type and BIB_ITEMS[bibkey]['year'] == cur_year %}
+                    {% set coverpng = BIB_ITEMS[bibkey]['coverpng'] %}
+                    {% set bibkey = BIB_ITEMS[bibkey]['coverpng'].replace('.png', '').lower() %}
+                        {% if BIB_ITEMS[bibkey]['cover_exists'] == 'True' %}
                         <div class="col-md-4 mb-4">
                             <div class="card">
-                                <a href="{{SITEURL}}/publications/{{ bibkey }}"><img class="card-img-top lazyload" data-src="{{SITEURL}}/{{ ('images/theses/' + coverpng) | resize_image('medium') }}" alt="{{ bib_items[bibkey]['title_thesis'] }}"></a>
+                                <a href="{{SITEURL}}/publications/{{ bibkey }}"><img class="card-img-top lazyload" data-src="{{SITEURL}}/{{ ('images/theses/' + coverpng) | resize_image('medium') }}" alt="{{ BIB_ITEMS[bibkey]['title_thesis'] }}"></a>
                             </div>
                         </div>
-
                         {% else %}
                             {% set _ = theses_without_cover.append(bibkey) %}
                         {% endif %}
@@ -70,12 +65,18 @@
                 {% endfor %}
             {% endfor %}
     {% endfor %}
-    <h3>Thesis (without cover)</h3>
-    <ul>
-    {% for tkey in theses_without_cover %}
-        <li>{{ bib_items[tkey]['title'] }} </li>
-    {% endfor %}
-    </ul>
     </div>
+
+
+    <!-- Thesis without cover -->
+    <div class="row">
+        <h3>Thesis (without cover)</h3>
+        <ul>
+        {% for bibkey in theses_without_cover %}
+            {% include "blocks/publication-item.html" %}
+        {% endfor %}
+        </ul>
+    </div>
+
     </div>
 {% endblock %}
