+++ radboudumc-theme/sass/theme.scss
@@ -62,3 +62,6 @@ Built on top of Bootstrap 4.
 // Util
 @import "icons";
 @import "video";
+
+// Year selection
+@import 'publications';
